% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatwave.r
\name{calculate_HWdays}
\alias{calculate_HWdays}
\title{calculate_HWdays}
\usage{
calculate_HWdays(data, clim_array, method = "EuroHeat", index = "ATI",
  wind = T, months = c(5:9), P_intesity = c(95), duration = 2,
  time_reference = "DAY")
}
\arguments{
\item{data.frame}{data Daily time series with colums' names: date(YYYY-MM-DD),tmed,tmax,tmin,rhum,vmed.}

\item{data.frame}{clim_array Name of location or the time series analized.}

\item{character}{method Method to assess critical and heatwave days. Default is "EuroHeat" and other are "TMAX","TMED","TMIN".}

\item{character}{index Biometerological index to calculate apparent temperature with ATI or UTCI. Default is "ATI" corresponding to the shaded version of ATI Steadman (2004).}

\item{logical}{wind If TRUE the wind in biometeorological index are considered.}

\item{numeric}{months The month her HW days are calculated. Default are c(5:9) corresponding to May-Dec.}

\item{numeric}{P_intesity If one of the "TMAX","TMED","TMIN" methods was selected is the level of intensity in term of percentiles.Default are c(95),}

\item{numeric}{duration. Minimum spell of critical days to be considered for heatave definition. Default is 2.}

\item{numeric}{time_reference. Temporal format of the referenced climatology: DAY (daily) or MON (monthly).Default is "DAY".}
}
\value{
list Return a list of data.frame respectively with monthly and daily values.
}
\description{
Calculate HWdays days and Critical ones adding some columns at data.frame.
Mean air temperature, mean relative humidity and extreme thermal values are required.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{heat}
\keyword{wave}
