% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatwave.r
\name{manage_GSOD}
\alias{manage_GSOD}
\title{manage_GSOD}
\usage{
manage_GSOD(id, name, ini_year = 1979, last_year = 2015, missing = NA,
  save_files = T)
}
\arguments{
\item{id}{character USAF code of weather station.}

\item{name}{character Name of location or ID of time series.}

\item{ini_year}{character Initial year of data time period, Default is "1979".}

\item{last_year}{character Final year of climatological reference time period, Default is "2016".}

\item{save_files}{logical If True save GSOD data file in RDS and CSV format. Default is F.}
}
\value{
Return  data.frame ready to use for heat-wave analitics.
}
\description{
Download and build a specific data.frame for heatwave analisys from a GSOD NOAA repository station based on its unique USAF code.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{GSOD}
